Started to implement word stemming
This service will work better if words are standardized. This can
be accomplished through the use of word stemming. Word stemming is
the process of algorithmiclly converting a word into a base-most
version without plurality or tense.

If words are standardized durring indexing and upon search, it will
allow for a greater number of relevent matches to be made.