Added user directory detection and word stemming
I realized a little late that I had previously hard coded the value
which is not particularly useful if I am trying to use this on other
machines. To that end, I added detection of the users home directory
using the $HOME variable.

I also added in word stemming, which allows for simplifying some
words into a common format. This will make it easier to identify
the correct document even if the query includes an unrelated plurality
or tense.