Added stemming to search
There was a small problem with the search function in which the input
was not being stemmed. This is not compatable with the index, which
stemmed everything.

Additionally, I fixed an issue in which the index script did not
split on all whitespace characters. This creates a scenario in which
there is an excess of bigram terms. This was fixed by using the
trim_whitespace() function as part of the basic String object.