Initial commit of Bayesian rewrite
I think that this tool may work better using bayseian inference to
predict the most relevent document. With this in mind, I have
started to tweak the interal components of this software to
utilize Bayes' Law for predicition. So far, I have revised the
corupus to write the term frequency for both the document and
the corpus to the index. Now all I need to do in adapt the search
function to calculate the posterior probability from the provided
conditional probabilities and a uniform prior probability.