Completed first major round of refactoring.
I've finished the first major round of refactoring. this has cleaned up
some of the Bayesian Classification code, simplified some of the sorting,
unified some calculations, and added some much needed documentation to the
processess at play.

I think that it may be interesting to create an abstracted class to handle
the classification of documents instead of it being the responsiblity of
the corpus. This different approach would allow for a cleaner means of
updating the search algorithim without having to wade through all of the
existing corpus and document code.

It would also be nice to clean up the interfaced used by the corpus and
document strucutres to make it a bit easier to work with the data.